amazonULC: A Data Package with Urban Land Cover Classifications for a Selection of Cities in the Brazilian Amazon

The amazonULC package aims to provide land cover classification maps of Brazilian Amazonian cities. Land cover maps can contribute to urban planning in Amazonian cities, with applications for monitoring urban sprawl, environmental protection areas, assisting in urban zoning, and identifying areas of high building density, among others. The amazonULC provides a urban land cover classification for the cities of Altamira, Cametá, Marabá, Santarém and part of the Metropolitan Area of Belém, all in the state of Pará, Brazilian Legal Amazon.

References

https://github.com/dias-bruno/amazonULC

See also

Author

Bruno Dos Santos, INPE bruno.santos@inpe.br