Urban Land Cover for the City of Cameta - Para State, Brazil (year of 2020)

A simple features data frame with the features and the land cover classification for Cameta. In order to obtain this classification, we used the mean-shift algorithm to segment an orthorectified image from the Wide Scan Multispectral and Panchromatic Camera (WPM) of the China-Brazil Earth Resources Satellite (CBERS-4A), dated 09/15/2020 (path: 212, row: 116). The WMP-CBERS image has 2 m spatial resolution for the panchromatic camera and 8 m for the multispectral camera. After segmentation, we extracted each segment's spectral, Grey-Level Co-occurrence Matrix texture measurements (GLCM), biophysical index, and geometry attributes. Then, after random and stratified sampling, we classified the segments with the Random Forest machine learning algorithm. The final classification has an accuracy of 77% to the testing samples (30% of the sample base). The classified area overlaps the urban limit of Cametá and is about 44 km². The data is in WGS84 Coordinate Systems, EPSG 4326. The polygons were classified into:"Shrub Vegetation" (SV), "Herbaceous Vegetation" (HV), "Water" (Wa), "Exposed Ground" (EG), "High Gloss Cover" (HG), "Ceramic Cover" (Ce), "Fiber Cement Cover" (FC), "Asphalt Road" (As), "Terrain Road" (Te), "Cloud" (Cl) and "Shadow" (Sh).

Usage

data(cameta_class)

Format

A simple features data frame with 47,834 rows and 34 variables:

ID

Unique identifier for each polygon in the data frame.

ASM_min

Minimum value of the Angular Second Moment (ASM, also called Uniformity). This is a measure of local homogeneity. Calculated over a moving window of 25x25 pixels, isotropically.

B0_standar

Standard Deviation of the blue band for all pixels inside the object.

B1_amplitu

Amplitude value of the green band for all pixels inside the object. The amplitude means the maximum pixel value minus the minimum pixel value.

B1_contr

Contrast of the green band for all pixels inside the object. It returns a measure of the intensity contrast between a pixel and its southeast neighbor over the object. Contrast is 0 for a constant object.

B1_homogen

Homogeneity of the green band for all pixels inside the object. It assumes higher values for smaller differences in the GLCM. Also called Inverse Difference Moment. Homogeneity is 1 for a diagonal GLCM.

B1_median

Median value of the green band for all pixels inside the object. The amplitude means the maximum pixel value minus the minimum pixel value.

B1_skewnes

Skewness value of the green band for all pixels inside the object.

B1_standar

Standard Deviation of the green band for all pixels inside the object.

B2_amplitu

Amplitude value of the red band for all pixels inside the object. The amplitude means the maximum pixel value minus the minimum pixel value.

B2_dissimi

Dissimilarity of the red band for all pixels inside the object. It measures how different the elements of the GLCM are from each other and it is high when the local region has a high contrast.

B2_homogen

Homogeneity of the red band for all pixels inside the object. It assumes higher values for smaller differences in the GLCM. Also called Inverse Difference Moment. Homogeneity is 1 for a diagonal GLCM.

B2_skewnes

Skewness value of the red band for all pixels inside the object.

B2_standar

Standard Deviation of the red band for all pixels inside the object.

B3_contr

Contrast of the near-infrared band for all pixels inside the object. It returns a measure of the intensity contrast between a pixel and its southeast neighbor over the object. Contrast is 0 for a constant object.

B3_mean

Average value of the near-infrared band for all pixels inside the object.

B3_skewnes

Skewness value of the near-infrared band for all pixels inside the object.

BAIN_mean

Average value of the Bare soil area index (near-infrared version): the normalized division of the blue band by the near-infrared band.

BAIN_std

Standard Deviation of the Bare soil area index (near-infrared version): the normalized division of the blue band by the near-infrared band.

BAI_mean

Average value of the Bare soil area index (red version): the normalized division of the blue band by the red band.

BAI_std

Standard Deviation of the Bare soil area index (red version): the normalized division of the blue band by the red band.

CONTR_min

Minimum value of the Constrast. This measure analyses the image contrast (locally gray-level variations) as the linear dependency of grey levels of neighboring pixels (similarity). Calculated over a moving window of 25x25 pixels, isotropically.

CONTR_std

Standard Deviation of the Constrast. This measure analyses the image contrast (locally gray-level variations) as the linear dependency of grey levels of neighboring pixels (similarity). Calculated over a moving window of 25x25 pixels, isotropically.

ENTR_max

Maximum value of the Entropy. This measure analyses the randomness. It is high when the values of the moving window have similar values. Calculated over a moving window of 25x25 pixels, isotropically.

ENTR_min

Minimum value of the Entropy. This measure analyses the randomness. It is high when the values of the moving window have similar values. Calculated over a moving window of 25x25 pixels, isotropically.

MOC1_mean

Average value of the Information Measures of Correlation. Calculated over a moving window of 25x25 pixels, isotropically.

PCIRCLE

Relates the areas of the object and the smallest circumscribing circle around the object.

PDENSITY

This feature corresponds to the ratio between the polygon area and the polygon radius.

PGYRATIUS

This feature is equals the average distance between each vertex of the polygon and it's centroid. The more similar to a circle is the object, the more likely the centroid will be inside it, and therefore this feature will be closer to 0.

P_PERARAT

Calculates the ratio between the perimeter and the area of an object.

P_PERIM

Returns the perimeter of the object.

TARGET

Variable related to the sampling process. Null value means that the feature was not selected during sampling.

CLASS

The land cover class in which the polygon belongs.

geometry

geometry of the simple features.

Source

Vrabel, J.C., Stensaas, G.L., Anderson, C., Christopherson, J., Kim, M., Park, S., and Cantrell, S., 2021, System characterization report on the China-Brazil Earth Resources Satellite-4A (CBERS–4A), chap. J of Ramaseri Chandra, S.N., comp., System characterization of Earth observation sensors: U.S. Geological Survey Open-File Report 2021–1030, 35 p., https://doi.org/10.3133/ofr20211030J.

dos Santos BD, de Pinho CMD, Oliveira GET, Korting TS, Escada MIS, Amaral S. Identifying Precarious Settlements and Urban Fabric Typologies Based on GEOBIA and Data Mining in Brazilian Amazon Cities. Remote Sensing. 2022; 14(3):704. https://doi.org/10.3390/rs14030704

Haralick, R.M., K. Shanmugam, and I. Dinstein (1973). Textural features for image classification. IEEE Transactions on Systems, Man, and Cybernetics, SMC-3(6):610-621.

Examples

 data(cameta_class)
#> Warning: file ‘cameta_class.rda’ has magic number 'versi'
#>   Use of save versions prior to 2 is deprecated
#> Error in load(file, envir = tmp_env): bad restore file magic number (file may be corrupted) -- no data loaded
 class <- cameta_class$CLASS
#> Error: object 'cameta_class' not found